week4/Day3/Q2/requirements.txt
flask==3.0.2
numpy==1.26.4
python-dotenv==1.0.1
openai==1.12.0 